+++ .github/workflows/check-plugin-version.yaml
@@ -3,7 +3,7 @@ name: TS Plugin Workers Package Version Check
 on:
   pull_request:
     paths:
-      - 'src/**'
+      - 'packages/plugin/src/**'
     branches: [main]
     types: [opened, synchronize, labeled, unlabeled]
   workflow_dispatch:
@@ -14,42 +14,7 @@ jobs:
     steps:
       - uses: actions/checkout@v2
 
-      - name: Check for "skip-version-check" label
-        id: check_label
-        shell: bash
-        run: |
-          SKIP_VERSION_BUMP=${{contains(github.event.pull_request.labels.*.name, 'skip-version-check') }}
-          if $SKIP_VERSION_BUMP; then
-            echo "Skip version check label found."
-          else
-            exit 1 
-          fi
-        continue-on-error: true
-
-      - name: Get base branch version
-        if: steps.check_label.outcome == 'failure'
-        shell: bash
-        run: |
-          git fetch origin ${{ github.base_ref }}:refs/remotes/origin/${{ github.base_ref }}
-          echo "BASE_VERSION=$(git show origin/${{ github.base_ref }}:package.json | node -p "require('./package.json').version")" >> $GITHUB_ENV
-
-      - name: Get PR branch version
-        if: steps.check_label.outcome == 'failure'
-        shell: bash
-        run: |
-          echo "PR_VERSION=$(node -p "require('./package.json').version")" >> $GITHUB_ENV
-
-      - name: Check if version is bumped
-        id: check_bump
-        if: steps.check_label.outcome == 'failure'
-        shell: bash
-        run: |
-          echo "Comparing base version $BASE_VERSION to the current version $PR_VERSION."
-          COMPARABLE_BASE_VERSION=$($GITHUB_WORKSPACE/scripts/ci/parse_version.sh $BASE_VERSION)
-          COMPARABLE_PR_VERSION=$($GITHUB_WORKSPACE/scripts/ci/parse_version.sh $PR_VERSION)
-          if $COMPARABLE_BASE_VERSION -lt $COMPARABLE_PR_VERSION; then
-            echo "Version bumped in package.json."
-          else
-            echo "Error: The version in package.json has not been bumped."
-            exit 1
-          fi
+      - name: Check version
+        uses: ./.github/actions/check-version
+        with:
+          project_path: packages/plugin
