--- .npmignore
@@ -1,3 +1,0 @@
-example/
-src/
-wasm/
