+++ base.prettierignore
@@ -1,0 +1,4 @@
+packages/
+node_modules/
+target/
+pnpm-lock.yaml
