+++ cargo.toml
@@ -1,3 +1,3 @@
 [workspace]
-members = ["wasm"]
+members = ["packages/wasm"]
 resolver = "2"
