--- example/.prettierrc
@@ -1,6 +1,0 @@
-{
-	"printWidth": 140,
-	"singleQuote": true,
-	"semi": true,
-	"useTabs": true
-}
