--- example/package.json
@@ -1,15 +1,0 @@
-{
-	"name": "example",
-	"version": "0.0.0",
-	"private": true,
-	"scripts": {
-		"deploy": "wrangler deploy",
-		"start": "wrangler dev"
-	},
-	"devDependencies": {
-		"@cloudflare/workers-types": "^4.20230419.0",
-		"@timsexperiments/ts-plugin-workers": "file:..",
-		"typescript": "^5.0.4",
-		"wrangler": "^3.0.0"
-	}
-}
