+++ packages/eslint/.gitignore
@@ -1,0 +1,4 @@
+.turbo/
+node_modules/
+
+index.d.ts
