+++ packages/eslint/package.json
@@ -1,0 +1,27 @@
+{
+  "name": "eslint-config",
+  "private": true,
+  "author": "timsexperiments <tim@timsexperiments.foo>",
+  "files": [
+    "ts-base.config.js"
+  ],
+  "types": "index.d.ts",
+  "scripts": {
+    "format:check": "prettier --check **/*.{js,json}"
+  },
+  "keywords": [],
+  "license": "ISC",
+  "dependencies": {
+    "@typescript-eslint/eslint-plugin": "^7.4.0",
+    "@typescript-eslint/parser": "^7.4.0",
+    "eslint-config-prettier": "^9.1.0"
+  },
+  "peerDependencies": {
+    "eslint-plugin-prettier": "^5.1.3"
+  },
+  "devDependencies": {
+    "@types/eslint": "^8.56.6",
+    "prettier": "^3.2.5"
+  },
+  "type": "module"
+}
