+++ packages/example/.editorconfig
