+++ packages/example/.prettierignore
@@ -1,0 +1,2 @@
+example/bindings.d.ts
+
