+++ packages/example/.vscode/settings.json
