+++ packages/plugin/.prettierignore
@@ -1,0 +1,3 @@
+dist/
+.turbo/
+node_modules/
