+++ packages/plugin/eslint.config.mjs
@@ -1,0 +1,4 @@
+import baseConfig from 'eslint-config/ts-base.config.js';
+
+/** @type {import('eslint').Linter.FlatConfig[]} */
+export default [...baseConfig];
