+++ packages/plugin/src/write_definitions.ts
