+++ packages/plugin/tsconfig.json
@@ -8,8 +8,7 @@
     "jsx": "react-jsx",
     "allowJs": true,
 
-    // Bundler mode
-    "moduleResolution": "Node",
+    "declaration": true,
     "sourceMap": true,
 
     // Best practices
@@ -25,5 +24,5 @@
     "esModuleInterop": true,
     "outDir": "dist"
   },
-  "include": ["src/**/*", "eslint.config.mjs"]
+  "include": ["src/**/*.ts"]
 }
