+++ packages/wasm/src/error.rs
